+++ packages/@baldeweg/ui/src/composables/useLocale.js
@@ -1,11 +1,13 @@
 import { ref, watchEffect } from 'vue'
 
+const globalLocale = ref(navigator.language)
+
 export function useLocale() {
-  const locale = ref(navigator.language)
+  const locale = ref(globalLocale.value)
 
   watchEffect(() => {
     document.documentElement.setAttribute('lang', locale.value)
   })
 
-  return { locale }
+  return { globalLocale, locale }
 }